Подправил detekt
