+++ app/src/main/java/ru/practicum/android/diploma/ui/search/model/TeamMemberUI.kt
@@ -1,0 +1,7 @@
+package ru.practicum.android.diploma.ui.search.model
+
+data class TeamMemberUI(
+    val name: String,
+    val role: String,
+    val iconResId: Int? = null
+)
