+++ app/src/main/res/values/strings.xml
@@ -36,4 +36,6 @@
     <string name="add_vacancy_to_favourite">Добавить в избранное</string>
     <string name="salary_for_sharing">Зарплата</string>
 
+    <string name="team_working_on_app">Над приложением работали</string>
+
 </resources>
